+++ src/channels/plugins/outbound/signal.ts
@@ -1,5 +1,8 @@
 import { sendMessageSignal } from "../../../../extensions/signal/src/send.js";
-import { resolveOutboundSendDep, type OutboundSendDeps } from "../../../infra/outbound/deliver.js";
+import {
+  resolveOutboundSendDep,
+  type OutboundSendDeps,
+} from "../../../infra/outbound/send-deps.js";
 import {
   createScopedChannelMediaMaxBytesResolver,
   createDirectTextMediaOutbound,
