+++ src/channels/plugins/whatsapp-shared.ts
@@ -1,4 +1,4 @@
-import { resolveOutboundSendDep } from "../../infra/outbound/deliver.js";
+import { resolveOutboundSendDep } from "../../infra/outbound/send-deps.js";
 import type { PluginRuntimeChannel } from "../../plugins/runtime/types-channel.js";
 import { escapeRegExp } from "../../utils.js";
 import { resolveWhatsAppOutboundTarget } from "../../whatsapp/resolve-outbound-target.js";
